created array of buttons, instead of individual declerations, also set max lenght 16 digits in textview in order to avoid overflow, also migrated to gradle
