--- app.project
@@ -1,1 +1,0 @@
-NAME= KasCalc
