+++ app/src/main/res/layout/activity_calculator.xml
