+++ app/src/main/res/menu/activity_calculator.xml
